feat: data class 별도 분리
