+++ app/src/main/java/com/eatssu/android/presentation/map/model/FavoritePartnership.kt
@@ -1,0 +1,7 @@
+package com.eatssu.android.presentation.map.model
+
+data class FavoritePartnership(
+    val storeName: String,
+    val category: String,
+    val period: String
+)
