+++ app/src/main/java/com/eatssu/android/presentation/map/model/MapRestaurantInfo.kt
@@ -1,0 +1,8 @@
+package com.eatssu.android.presentation.map.model
+
+data class MapRestaurantInfo(
+    val collegeName: String?,
+    val departmentName: String?,
+    val period: String,
+    val benefit: String
+)
